-use 2L and 3L clients to make requests - access tokens are encapsulated inside the client instance - refactor auth.js to not include access token handling - async-await requests
