+++ routes/auth.js
@@ -1,124 +1,42 @@
 const express = require('express')
 
 const router = express.Router()
-const axios = require('axios')
-const qs = require('qs')
 
-const {
-  authorizationUrl,
-  tokenEndpoint,
-  userInfoEndpoint,
-  redirectUri,
-  base64Credentials,
-} = require('../helpers/spotifyOauth')
+const SpotifyClient = require('../services/spotify')
 
-// cached access token
-// IMPROVEMENT: put in a class instead
-// IMPROVEMENT: use a better cache or DB to persist between server reloads
-let savedAccessTokenPayload = null
 let userInfo
 let message
 
-function getAndRenderUserProfile(tokenObj, res) {
-  axios({
-    method: 'GET',
-    url: userInfoEndpoint,
-    headers: {
-      Authorization: `Bearer ${tokenObj.access_token}`,
-      Accept: 'application/json',
-    },
-  }).then((result) => {
-    // console.log(result.data)
-    userInfo = result.data
+const threeLeggedClient = SpotifyClient.getThreeLeggedClient()
+const readUserAuthorizationUrl = threeLeggedClient.getAuthorizationUrl({
+  scope: SpotifyClient.getScope('USER_READ'),
+  state: 'foo=bar&fizz=buzz', // app-specific state that we want to persist on callback
+})
 
-    // res.redirect('user') // -> auth2/user
-    res.redirect('/') // -> auth2/user
-    // res.redirect('index')
-  })
-}
-
-router.get('/', (req, res) => {
+router.get('/', async (req, res) => {
   // views/index.ejs gets assets from public/ folder
-
   res.render('index', {
     message,
     user: userInfo,
-    auth_url: authorizationUrl,
+    auth_url: readUserAuthorizationUrl,
   })
   // res.sendFile(__dirname + '/public/index.html');
 })
 
-// LOGIN is now managed by frontend
-
-// router.get('/login', (req, res) => {
-//   // if token exists already, do the request
-//   if (savedAccessTokenPayload) {
-//     message = 'Using cached token 🎁'
-//     console.log(message)
-//     getAndRenderUserProfile(savedAccessTokenPayload, res)
-//   } else {
-//     // STEP 1: direct user to Spotify login page
-//     message = ''
-//     res.render('login', {
-//       auth_url: authorizationUrl,
-//     })
-//   }
-// })
-
-// STEP 2
-router.get('/callback', (req, res) => {
-  console.log(req.query)
-
-  // get code from query string
-  const authCode = req.query.code
-
-  // STEP 2: send a request to exchange auth code with access token
-
-  axios({
-    method: 'post',
-    url: tokenEndpoint,
-    data: qs.stringify({
-      grant_type: 'authorization_code',
-      code: authCode,
-      redirect_uri: redirectUri,
-    }),
-    headers: {
-      Authorization: `Basic ${base64Credentials}`,
-    },
+router.get('/callback', async (req, res) => {
+  await threeLeggedClient.completeThreeLeggedAuth({
+    authCode: req.query.code, // auth code in callback
   })
-    .then((result) => {
-      console.log(result.data)
-      // cache entire response payload
-      savedAccessTokenPayload = result.data
 
-      // STEP 3: send a request using access_token
-      getAndRenderUserProfile(savedAccessTokenPayload, res)
-    })
-    .catch((err) => {
-      console.error(err.response.data)
-    })
+  // 3L client should have an access token now
+  userInfo = await threeLeggedClient.getMe()
+  res.redirect('/')
 })
 
-// router.get('/user', (req, res) => {
-//   if (userInfo) {
-//     res.render('index', {
-//       user: userInfo,
-//       message,
-//       auth_url: authorizationUrl,
-//     })
-//   } else {
-//     // res.redirect('/login')
-//     res.redirect('/')
-//   }
-// })
-
 router.get('/logout', (req, res) => {
-  // delete saved access token, thus
-  // restarting Auth code flow
-  savedAccessTokenPayload = null
+  threeLeggedClient.logout()
   userInfo = null
 
-  // res.redirect('/login')
   res.redirect('/')
 })
